reorganize helper code
