+++ provider/pkg/provider/provider_test.go
@@ -15,12 +15,28 @@
 package provider
 
 import (
+	"context"
+	"os"
 	"testing"
 
+	"fmt"
+	"log"
+
+	pbempty "github.com/golang/protobuf/ptypes/empty"
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/types"
+	gomegatypes "github.com/onsi/gomega/types"
 	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/kinds"
+	"github.com/pulumi/pulumi/pkg/v3/resource/provider"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
+	"github.com/pulumi/pulumi/sdk/v3/go/common/util/rpcutil"
+	pulumirpc "github.com/pulumi/pulumi/sdk/v3/proto/go"
 	"github.com/stretchr/testify/assert"
+	"google.golang.org/grpc"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/client-go/tools/clientcmd"
+	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
 )
 
 var (
@@ -229,4 +245,100 @@ func Test_isListURN(t *testing.T) {
 			assert.Equalf(t, tt.want, kinds.IsListURN(tt.args.urn), "isListURN(%v)", tt.args.urn)
 		})
 	}
+}
+
+// CheckFailure matches a CheckFailure by property and reason.
+func CheckFailure(prop string, reason types.GomegaMatcher) gomegatypes.GomegaMatcher {
+	return And(
+		WithTransform(func(failure *pulumirpc.CheckFailure) string {
+			return failure.GetProperty()
+		}, Equal(prop)),
+		WithTransform(func(failure *pulumirpc.CheckFailure) string {
+			return failure.GetReason()
+		}, reason))
+}
+
+// KubeconfigAsString converts a Kubernetes configuration to a string.
+func KubeconfigAsString(config *clientcmdapi.Config) string {
+	contents, err := clientcmd.Write(*config)
+	Expect(err).ToNot(HaveOccurred())
+	return string(contents)
+}
+
+// KubeconfigAsFile converts a Kubernetes configuration to a string.
+func KubeconfigAsFile(config *clientcmdapi.Config) string {
+	f, _ := os.CreateTemp("", "kubeconfig-")
+	f.Close()
+	DeferCleanup(func() {
+		os.Remove(f.Name())
+	})
+	err := clientcmd.WriteToFile(*config, f.Name())
+	Expect(err).ToNot(HaveOccurred())
+	return f.Name()
+}
+
+// A mock engine for test purposes.
+type mockEngine struct {
+	pulumirpc.UnsafeEngineServer
+	t            TestingTB
+	logger       *log.Logger
+	rootResource string
+}
+
+// Log logs a global message in the engine, including errors and warnings.
+func (m *mockEngine) Log(ctx context.Context, in *pulumirpc.LogRequest) (*pbempty.Empty, error) {
+	m.t.Logf("%s: %s", in.GetSeverity(), in.GetMessage())
+	if m.logger != nil {
+		m.logger.Printf("%s: %s", in.GetSeverity(), in.GetMessage())
+	}
+	return &pbempty.Empty{}, nil
+}
+
+// GetRootResource gets the URN of the root resource, the resource that should be the root of all
+// otherwise-unparented resources.
+func (m *mockEngine) GetRootResource(ctx context.Context, in *pulumirpc.GetRootResourceRequest) (*pulumirpc.GetRootResourceResponse, error) {
+	return &pulumirpc.GetRootResourceResponse{
+		Urn: m.rootResource,
+	}, nil
+}
+
+// SetRootResource sets the URN of the root resource.
+func (m *mockEngine) SetRootResource(ctx context.Context, in *pulumirpc.SetRootResourceRequest) (*pulumirpc.SetRootResourceResponse, error) {
+	m.rootResource = in.GetUrn()
+	return &pulumirpc.SetRootResourceResponse{}, nil
+}
+
+// newMockHost creates a mock host for test purposes and returns a client.
+// Dispatches Engine RPC calls to the given engine.
+func newMockHost(ctx context.Context, engine pulumirpc.EngineServer) *provider.HostClient {
+	cancel := make(chan bool)
+	go func() {
+		<-ctx.Done()
+		close(cancel)
+	}()
+	handle, err := rpcutil.ServeWithOptions(rpcutil.ServeOptions{
+		Cancel: cancel,
+		Init: func(srv *grpc.Server) error {
+			pulumirpc.RegisterEngineServer(srv, engine)
+			return nil
+		},
+		Options: rpcutil.OpenTracingServerInterceptorOptions(nil),
+	})
+	if err != nil {
+		panic(fmt.Errorf("could not start host engine service: %v", err))
+	}
+
+	go func() {
+		err := <-handle.Done
+		if err != nil {
+			panic(fmt.Errorf("host engine service failed: %v", err))
+		}
+	}()
+
+	address := fmt.Sprintf("127.0.0.1:%v", handle.Port)
+	hostClient, err := provider.NewHostClient(address)
+	if err != nil {
+		panic(fmt.Errorf("could not connect to host engine service: %v", err))
+	}
+	return hostClient
 }
